Akt-backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
dnspython
python-dotenv